Copy Perl modules with non-mainline versions
For stand-alone builds, the pm portion of the compiled xs files
needs to match versions.  There are a few instances where there
are Perl module versions that are different from the mainline
squeeze server.  For those instances, the .pm files are copied
to the arch/5.X directory after the module is built.