feat(ICU): Upgrade ICU to 58.2, necessary collation and build patches.
This upgrades the ICU library and associated collation files to ICU 58.2, as
well as provides repeatable instructions for creation of the collation file with
arbitrary ICU versions. Additionally, this commit adds the necessary patches for
ICU to build with FreeBSD, tested on 7.2 and 11.1. Several code corrections were
made to correctly identify the compiler in use, warn of a compiler mismatch
between buildme.sh and perl, halt for incompatible compiler versions or unkonwn
compiler types. ICU 58.2 requires std=c99 support, so this is the primary factor
to consider when evaluating compiler compatibility. The IO::Interface test was
disabled if the build is being conducted in a FreeBSD jail, because the test
requires 1o0 to be 127.0.0.1. Lastly, a DBD::SQLite + ICU build patch was added
to enable build when the c++ library is -lc++ (instead of -lstdc++). The tests
to determine which c++ library is in use (and thus control patch application)
were also added to buildme.sh.