DBI & DBD::SQLite - Tidy up buildme.sh & re-enable tests
1) DBI is being built twice by ’build_all’

Once directly, and then again when DBD::SQLite is built. A separate build incantation is
used in each case.

This change stops the 'build_all' function from building DBI, passing the responsibility
directly over to the DBD::SQLite build process.

Note that DBD::Sqlite references DBI during its own build process, so it is desirable to
build DBI first. That avoids any issue that may arise should DBD::SQLite be built against
a version of DBI that is already installed on the system. A reminder has been added to
this effect.

2) Re-enable testing on DBI-1.616 (installed for Perl <= 5.16).

Testing was turned off in 2017. One very small element of:
‘Upgrade ICU to 58.2, necessary collation and build patches’
https://github.com/Logitech/slimserver-vendor/commit/cb7e0a

This appears to have replicated the approach adopted by the then existing DBD::SQLite
build of DBI. But testing wasn’t needed there.

3) Re-enable testing on DBD::SQLite

This was originally turned off for Perl 5.16 in 2013. Refer:
‘Add support for building modules for Perl 5.16’
https://github.com/Logitech/slimserver-vendor/commit/6b67d4

However, tests are now known to succeed on:
  Debian Perl 5.24
  OSX 10.12 - The system Perl 5.16 & 5.18, and Homebrew Perl 5.26.

Therefore re-enabling tests on all Perls.