Quieten tar - 'spams' STDERR on MacOS
At present, 'tar' is being called with the verbose flag when archives
are unpacked. This generates a listing of the files as they are
unpacked, which is not particularly useful.

GNU tar outputs the listing to STDOUT. The function 'tar_wrapper' then
throws that straight into /dev/null, so it is not seen.

MacOS tar (BSDTAR) outputs the listing to STDERR. This makes for a very
noisy STDERR and interferes with the normal flow of build error/warning
messages.

This proposed change quietens matters by removing the verbose flag from
tar invocations. tar_wrapper's STDOUT redirection then becomes
unnecessary, so it is removed.